Fix failing tests that outlived their production code
- test_merge_artifacts: "measurements" is now a known-integer count
  column forcibly cast to Int64 during merge, so mismatches no longer
  promote to Float64.
- test_sbed_thinning: the locator's n_candidates constructor arg was
  replaced by candidate_step_hz-based thinning; pass an explicit
  candidate_step_hz so the 20 MHz test domain still thins to <=100
  candidates.
- test_smc_focus_window: narrowing is now gated behind
  NVISION_MIN_STEPS_BEFORE_NARROWING (default 8), so tests calling
  _resample() directly need _step_count bumped past the gate. The two
  shoulder-based tests exercised raw-observation dip detection that was
  replaced by particle/posterior-driven "Active-Range Union Focusing";
  rewrote them to drive that mechanism with realistic converged
  particle distributions and seeded the RNG since resample's rejuvenation
  step makes window width sensitive to draw.

+++ tests/test_merge_artifacts.py
@@ -42,10 +42,12 @@ def test_merge_locator_results_schema_mismatch(tmp_path: Path):
     # Execute merge
     merged_df = merge_locator_results_with_existing(new_df, out_dir, log)
 
-    # Verify that they merged successfully and measurements column is aligned (Float64)
+    # Verify that they merged successfully and measurements column is aligned.
+    # 'measurements' is a known-integer count column, so mismatches are
+    # normalized to Int64 rather than promoted to Float64.
     assert len(merged_df) == 1
-    assert merged_df.schema["measurements"] == pl.Float64
-    assert merged_df.get_column("measurements")[0] == 89.0
+    assert merged_df.schema["measurements"] == pl.Int64
+    assert merged_df.get_column("measurements")[0] == 89
 
 
 def test_merge_locator_results_keeps_other_combinations_even_with_no_cache(tmp_path: Path):
